feat(visits): enhance visit management with new endpoints and client-specific queries
