+++ src/routes/visits.admin.route.js
@@ -1,13 +1,14 @@
 import { Router } from "express";
-import { createVisit, getAllVisitsCount, getCancelledVisits, getCompletedVisits, getConfirmedVisits, getPastVisits, getPendingVisits, getPendingVisitsCount, getUpcomingVisits, updateVisit, updateVisitStaff } from "../controller/visits.admin.controller.js";
+import { createVisit, getAllVisitsCount, getCancelledVisits, getCompletedVisits, getConfirmedVisits, getConfirmedVisitsCount, getPastVisits, getPendingVisits, getPendingVisitsCount, getUpcomingVisits, updateVisit, updateVisitStaff } from "../controller/visits.admin.controller.js";
 import { verifyJWT } from "../middleware/auth.middleware.js";
 import { isAdmin } from "../middleware/role.middleware.js";
 
 const router = Router();
 
 router.post("/create-visit", verifyJWT, isAdmin, createVisit);
-router.get('/get-all-visits-count', verifyJWT, isAdmin, getAllVisitsCount);
-router.get('/get-pending-visits-count', verifyJWT, isAdmin, getPendingVisitsCount)
+router.get('/get-all-visits-count/:client', verifyJWT, isAdmin, getAllVisitsCount);
+router.get('/get-pending-visits-count/:client', verifyJWT, isAdmin, getPendingVisitsCount)
+router.get('/get-confirmed-visits-count/:client', verifyJWT, isAdmin, getConfirmedVisitsCount)
 router.get('/get-confirmed-visits/:client', verifyJWT, isAdmin, getConfirmedVisits);
 router.get('/get-pending-visits/:client', verifyJWT, isAdmin, getPendingVisits);
 router.get('/get-completed-visits/:client', verifyJWT, isAdmin, getCompletedVisits);
