+++ src/routes/visits.client.route.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { createVisit, getCancelledVisits, getConfirmedVisits, getPendingVisits, getCompletedVisits, getVisitById, updateVisit, getNextVisit, getPastVisits, getUpcomingVisits } from "../controller/visits.client.controller.js";
+import { createVisit, getCancelledVisits, getConfirmedVisits, getPendingVisits, getCompletedVisits, updateVisit, getNextVisit, getPastVisits, getUpcomingVisits } from "../controller/visits.client.controller.js";
 import { verifyJWT } from "../middleware/auth.middleware.js";
 import { isClient } from "../middleware/role.middleware.js";
 
@@ -12,7 +12,6 @@ router.get('/get-completed-visits', verifyJWT, isClient, getCompletedVisits);
 router.get('/get-cancelled-visits', verifyJWT, isClient, getCancelledVisits)
 router.get('/get-past-visits', verifyJWT, isClient, getPastVisits)
 router.get('/get-upcoming-visits', verifyJWT, isClient, getUpcomingVisits)
-router.get('/get-visit/:id', verifyJWT, isClient, getVisitById)
 router.get('/get-next-visit', verifyJWT, isClient, getNextVisit)
 router.patch('/update-visit/:id', verifyJWT, isClient, updateVisit);
